--- babel.config.js
@@ -1,18 +1,0 @@
-module.exports = {
-  presets: [
-    [
-      "@babel/preset-env",
-      {
-        targets: {
-          chrome: "79",
-          ie: "11" // npm i regenerator-runtime 필요함
-        },
-        useBuiltIns: "usage", // 폴리필 사용 방식 지정
-        corejs: {
-          // 폴리필 버전 지정
-          version: 2
-        }
-      }
-    ]
-  ]
-};
