+++ index.html
@@ -3,7 +3,8 @@
   <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>검색<%= env %></title>
+    <title>검색</title>
+    <link rel="stylesheet" href="src/main.css">
   </head>
   <body>
     <div id="app">
@@ -26,6 +27,6 @@
       </div>
     </div>
 
-    <script type="text/javascript" src="axios.min.js"></script>
+    <!-- TODO: 웹팩으로 빌드한 자바스크립트를 여기에 로딩하세요 -->
   </body>
 </html>
