--- mocks/api/history/GET.json
@@ -1,4 +1,0 @@
-[
-  { "keyword": "검색이력 1", "date": "2019-12-02" }, 
-  { "keyword": "검색이력 2", "date": "2019-12-01" }
-]
