--- mocks/api/keywords/GET.json
@@ -1,6 +1,0 @@
-[
-  { "keyword": "이탈리아" }, 
-  { "keyword": "세프의요리" }, 
-  { "keyword": "제철" }, 
-  { "keyword": "홈파티 " }
-]
