--- mocks/api/search/GET.json
@@ -1,12 +1,0 @@
-[
-  {
-    "id": 1,
-    "name": "[햅버거] 수제버거",
-    "image": "https://images.unsplash.com/photo-1549611016-3a70d82b5040?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1432&q=80"
-  },
-  {
-    "id": 2,
-    "name": "[키친르쎌] 이탈리아 파티 세트 3~4인분",
-    "image": "https://cdn.bmf.kr/_data/product/H503E/300d931e3b8252ed628b6a3c2f56936b.jpg"
-  }
-]
