--- package.json
@@ -1,53 +1,0 @@
-{
-  "name": "lecture-frontend-dev-env",
-  "version": "1.0.0",
-  "description": "\"프론트엔드 개발 환경의 이해\" 강의 자료입니다.",
-  "main": "index.js",
-  "devDependencies": {
-    "@babel/core": "^7.7.7",
-    "@babel/preset-env": "^7.7.7",
-    "babel-loader": "^8.0.6",
-    "clean-webpack-plugin": "^3.0.0",
-    "connect-api-mocker": "^1.7.0",
-    "copy-webpack-plugin": "^5.1.1",
-    "core-js": "^2.6.11",
-    "css-loader": "^3.4.1",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.9.0",
-    "eslint-plugin-prettier": "^3.1.2",
-    "file-loader": "^5.0.2",
-    "html-webpack-plugin": "^3.2.0",
-    "mini-css-extract-plugin": "^0.9.0",
-    "node-sass": "^4.13.0",
-    "optimize-css-assets-webpack-plugin": "^5.0.3",
-    "prettier": "^1.19.1",
-    "regenerator-runtime": "^0.13.3",
-    "sass-loader": "^8.0.0",
-    "style-loader": "^1.1.2",
-    "terser-webpack-plugin": "^2.3.2",
-    "url-loader": "^3.0.0",
-    "webpack": "^4.41.5",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.10.1"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack-dev-server --progress",
-    "build": "NODE_ENV=production webpack --progress",
-    "lint": "eslint src --fix"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jeonghwan-kim/lecture-frontend-dev-env.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/jeonghwan-kim/lecture-frontend-dev-env/issues"
-  },
-  "homepage": "https://github.com/jeonghwan-kim/lecture-frontend-dev-env#readme",
-  "dependencies": {
-    "axios": "^0.19.1"
-  }
-}
