+++ src/app.js
@@ -1,5 +1,4 @@
 import MainController from "./controllers/MainController.js";
-import "./app.scss";
 
 document.addEventListener("DOMContentLoaded", () => {
   new MainController();
