+++ src/main.css
@@ -1,0 +1,396 @@
+/*! 빌드 날짜: 2020. 1. 11. 오전 11:11:06 */
+.FormView {
+  position: relative
+}
+
+.FormView input[type=text] {
+  display: block;
+  box-sizing: border-box;
+  width: 100%;
+  margin: 15px 0;
+  padding: 10px 15px;
+  font-size: 14px;
+  line-height: 1.5;
+  border: 1px solid #ccc
+}
+
+.FormView .btn-reset {
+  position: absolute;
+  top: 12px;
+  right: 10px;
+  border: none;
+  border-radius: 50%;
+  padding: 0;
+  width: 20px;
+  height: 20px
+}
+
+.FormView .btn-reset:before {
+  content: " ";
+  display: inline-block;
+  height: 20px;
+  width: 20px;
+  background: url(images/times-circle.png?2a32eee45a9f280f743fb36b7c530d98) no-repeat 50%/contain;
+  opacity: .15
+}
+
+.FormView .btn-reset:hover {
+  cursor: pointer
+}
+
+.ResultView ul {
+  margin: 0;
+  padding: 0
+}
+
+.ResultView ul li {
+  display: flex;
+  margin-bottom: 4px;
+  padding: 0 4px
+}
+
+.ResultView ul li:first-child {
+  padding-top: 4px
+}
+
+.ResultView ul li:last-child {
+  padding-bottom: 4px;
+  margin-bottom: 0
+}
+
+.ResultView ul li:hover {
+  cursor: pointer
+}
+
+.ResultView ul img {
+  max-width: 150px;
+  margin-right: 8px
+}
+
+.ResultView ul p {
+  display: flex;
+  align-items: center
+}
+
+.TabView {
+  display: flex;
+  margin: 0;
+  padding: 0
+}
+
+.TabView li {
+  display: inline-block;
+  width: 50%;
+  padding: 15px;
+  text-align: center;
+  box-sizing: border-box;
+  border-bottom: 1px solid #ddd;
+  background-color: #fff;
+  color: #aaa
+}
+
+.TabView li:hover {
+  cursor: pointer
+}
+
+.TabView li.active {
+  background-color: #2ac1bc;
+  color: #fff
+}
+
+.TabView li.active:hover {
+  cursor: auto
+}
+
+.KeywordView {
+  margin: 0;
+  padding: 0
+}
+
+.KeywordView li {
+  box-sizing: border-box;
+  display: block;
+  padding: 15px;
+  border-bottom: 1px solid #ccc;
+  position: relative
+}
+
+.KeywordView li:hover {
+  background-color: #effbfb;
+  cursor: pointer
+}
+
+.KeywordView li:last-child {
+  border-bottom: none
+}
+
+.KeywordView li .date,
+.KeywordView li .number {
+  margin-right: 15px;
+  color: #ccc
+}
+
+.KeywordView li .date {
+  position: absolute;
+  right: 50px;
+  top: 15px
+}
+
+.HistoryView {
+  margin: 0;
+  padding: 0
+}
+
+.HistoryView li {
+  box-sizing: border-box;
+  display: block;
+  padding: 15px;
+  border-bottom: 1px solid #ccc;
+  position: relative
+}
+
+.HistoryView li:hover {
+  background-color: #effbfb;
+  cursor: pointer
+}
+
+.HistoryView li:last-child {
+  border-bottom: none
+}
+
+.HistoryView li .date,
+.HistoryView li .number {
+  margin-right: 15px;
+  color: #ccc
+}
+
+.HistoryView li .date {
+  position: absolute;
+  right: 50px;
+  top: 15px
+}
+
+.HistoryView .btn-remove {
+  position: absolute;
+  right: 0;
+  top: 12px;
+  margin-right: 15px;
+  border: none;
+  border-radius: 50%;
+  padding: 0;
+  width: 20px;
+  height: 20px
+}
+
+.HistoryView .btn-remove:before {
+  content: " ";
+  display: inline-block;
+  height: 20px;
+  width: 20px;
+  background: url(images/times-circle.png?2a32eee45a9f280f743fb36b7c530d98) no-repeat 50%/contain;
+  opacity: .15
+}
+
+.HistoryView .btn-remove:hover {
+  cursor: pointer
+}
+
+/*! normalize.css v8.0.1 | MIT License | github.com/necolas/normalize.css */
+html {
+  line-height: 1.15;
+  -webkit-text-size-adjust: 100%
+}
+
+body {
+  margin: 0
+}
+
+main {
+  display: block
+}
+
+h1 {
+  font-size: 2em;
+  margin: .67em 0
+}
+
+hr {
+  box-sizing: content-box;
+  height: 0;
+  overflow: visible
+}
+
+pre {
+  font-family: monospace, monospace;
+  font-size: 1em
+}
+
+a {
+  background-color: transparent
+}
+
+abbr[title] {
+  border-bottom: none;
+  text-decoration: underline;
+  text-decoration: underline dotted
+}
+
+b,
+strong {
+  font-weight: bolder
+}
+
+code,
+kbd,
+samp {
+  font-family: monospace, monospace;
+  font-size: 1em
+}
+
+small {
+  font-size: 80%
+}
+
+sub,
+sup {
+  font-size: 75%;
+  line-height: 0;
+  position: relative;
+  vertical-align: baseline
+}
+
+sub {
+  bottom: -.25em
+}
+
+sup {
+  top: -.5em
+}
+
+img {
+  border-style: none
+}
+
+button,
+input,
+optgroup,
+select,
+textarea {
+  font-family: inherit;
+  font-size: 100%;
+  line-height: 1.15;
+  margin: 0
+}
+
+button,
+input {
+  overflow: visible
+}
+
+button,
+select {
+  text-transform: none
+}
+
+[type=button],
+[type=reset],
+[type=submit],
+button {
+  -webkit-appearance: button
+}
+
+[type=button]::-moz-focus-inner,
+[type=reset]::-moz-focus-inner,
+[type=submit]::-moz-focus-inner,
+button::-moz-focus-inner {
+  border-style: none;
+  padding: 0
+}
+
+[type=button]:-moz-focusring,
+[type=reset]:-moz-focusring,
+[type=submit]:-moz-focusring,
+button:-moz-focusring {
+  outline: 1px dotted ButtonText
+}
+
+fieldset {
+  padding: .35em .75em .625em
+}
+
+legend {
+  box-sizing: border-box;
+  color: inherit;
+  display: table;
+  max-width: 100%;
+  padding: 0;
+  white-space: normal
+}
+
+progress {
+  vertical-align: baseline
+}
+
+textarea {
+  overflow: auto
+}
+
+[type=checkbox],
+[type=radio] {
+  box-sizing: border-box;
+  padding: 0
+}
+
+[type=number]::-webkit-inner-spin-button,
+[type=number]::-webkit-outer-spin-button {
+  height: auto
+}
+
+[type=search] {
+  -webkit-appearance: textfield;
+  outline-offset: -2px
+}
+
+[type=search]::-webkit-search-decoration {
+  -webkit-appearance: none
+}
+
+::-webkit-file-upload-button {
+  -webkit-appearance: button;
+  font: inherit
+}
+
+details {
+  display: block
+}
+
+summary {
+  display: list-item
+}
+
+[hidden],
+template {
+  display: none
+}
+
+#app {
+  min-width: 320px;
+  max-width: 768px;
+  margin-left: auto;
+  margin-right: auto
+}
+
+.container {
+  margin-left: 8px;
+  margin-right: 8px
+}
+
+header {
+  border-bottom: 1px solid #ddd;
+  margin: 0;
+  text-align: center
+}
+
+.content {
+  border: 1px solid #ddd
+}
