--- src/models/request.js
@@ -1,12 +1,0 @@
-import axios from "axios";
-
-export default async function request(method, url, data, headers) {
-  const result = await axios({
-    method,
-    url,
-    data,
-    headers
-  });
-
-  return result.data;
-}
