--- src/scss/mixins.scss
@@ -1,52 +1,0 @@
-@import './variables.scss';
-
-@mixin btn-close {
-  $width: 20px;
-  border: none;
-  border-radius: 50%;
-  padding: 0;
-  width: $width;
-  height: $width;
-  &:before {
-    content: ' ';
-    display: inline-block;
-    height: $width;
-    width: $width;
-    background: url('../images/times-circle.png') no-repeat center / contain;
-    opacity: .15;
-  }
-  &:hover {
-    cursor: pointer;
-  }
-}
-
-@mixin list {
-  margin: 0;
-  padding: 0;
-  li {
-    box-sizing: border-box;
-    display: block;
-    padding: 15px;
-    border-bottom: 1px solid #ccc;
-    position: relative;
-
-    &:hover {
-      background-color: lighten(#2ac1bc, 50);
-      cursor: pointer;
-    }
-    &:last-child {
-      border-bottom: none;
-    }
-    .number {
-      margin-right: 15px;
-      color: #ccc;
-    }
-    .date {
-      position: absolute;
-      right: 50px;
-      top: 15px;
-      margin-right: 15px;
-      color: #ccc;
-    }
-  }
-}
