--- src/scss/variables.scss
@@ -1,9 +1,0 @@
-$brand-color: #2ac1bc;
-$primary-color: blue;
-$secondary-color: #aaa;
-$white-color: #fff;
-
-$border-color: lighten($secondary-color, 20);
-
-$span: 4px;
-
