+++ src/views/FormView.js
@@ -1,5 +1,4 @@
 import View from "./View.js";
-import "./FormView.scss";
 
 export default class FormView extends View {
   constructor(el) {
