--- src/views/FormView.scss
@@ -1,22 +1,0 @@
-@import '../scss/mixins.scss';
-
-
-.FormView {
-  position: relative;
-  input[type=text] {
-    display: block;
-    box-sizing: border-box;
-    width: 100%;
-    margin: 15px 0 15px 0;
-    padding: 10px 15px;
-    font-size: 14px;
-    line-height: 1.5;
-    border: 1px solid #cccccc;
-  }
-    .btn-reset {
-    position: absolute;
-    top: 12px;
-    right: 10px;
-    @include btn-close();
-  }
-}
