+++ src/views/HistoryView.js
@@ -1,5 +1,4 @@
 import KeywordView from "./KeywordView.js";
-import "./HistoryView.scss";
 
 export default class HistoryView extends KeywordView {
   constructor(el) {
