+++ src/views/KeywordView.js
@@ -1,5 +1,4 @@
 import View from "./View.js";
-import "./KeywordView.scss";
 
 export default class KeywordView extends View {
   constructor(el) {
