--- src/views/KeywordView.scss
@@ -1,5 +1,0 @@
-@import '../scss/mixins.scss';
-
-.KeywordView {
-  @include list();
-}
