+++ src/views/TabView.js
@@ -1,6 +1,5 @@
 import View from "./View.js";
 
-import "./TabView.scss";
 
 export default class TabView extends View {
   constructor(el) {
